Move structs to their proper files.
